@@ -12,7 +12,7 @@ keywords = ["pattern-matching", "cybersecurity", "forensics", "malware", "yara"]
 #   .github/workflows/tests.yaml
 #   .github/workflows/code_health.yaml
 #
-rust-version = "1.74.0"
+rust-version = "1.76.0"
 
 [workspace]
 members = [
@@ -75,8 +75,6 @@ num-traits = "0.2.19"
 num-derive = "0.4.2"
 p256 = "0.13.2"
 p384 = "0.13.0"
-pest = "2.7.10"
-pest_derive = "2.7.10"
 pretty_assertions = "1.4.0"
 protobuf = "3.5.0"
 protobuf-codegen = "3.5.0"
@@ -99,8 +97,8 @@ thiserror = "1.0.61"
 # issue: https://github.com/1crcbl/tlsh-rs/issues/2.
 tlsh-fixed = "0.1.1"
 uuid = "1.4.1"
-walrus = "0.20.1"
-wasmtime = "19.0.2"
+walrus = "0.21.1"
+wasmtime = "22.0.0"
 x509-parser = "0.16.0"
 yaml-rust = "0.4.5"
 yansi = "1.0.1"
