@@ -1,7 +1,9 @@
 /*! Functions for converting a hex pattern AST into a HIR. */
 
+use crate::CompileError;
 use regex_syntax::hir;
 use yara_x_parser::ast;
+use yara_x_parser::ast::WithSpan;
 
 use crate::compiler::context::CompileContext;
 use crate::compiler::warnings::Warning;
@@ -10,7 +12,7 @@ use crate::compiler::ByteMaskCombinator;
 pub(in crate::compiler) fn hex_pattern_hir_from_ast(
     ctx: &mut CompileContext,
     pattern: &ast::HexPattern,
-) -> hir::Hir {
+) -> Result<hir::Hir, Box<CompileError>> {
     hex_tokens_hir_from_ast(ctx, &pattern.identifier, &pattern.tokens)
 }
 
@@ -18,16 +20,27 @@ fn hex_tokens_hir_from_ast(
     ctx: &mut CompileContext,
     pattern_ident: &ast::Ident,
     tokens: &ast::HexTokens,
-) -> hir::Hir {
+) -> Result<hir::Hir, Box<CompileError>> {
     let mut hir_tokens = Vec::with_capacity(tokens.tokens.len());
-    let ast_tokens = tokens.tokens.iter();
+    let mut ast_tokens = tokens.tokens.iter().peekable();
 
-    for token in ast_tokens {
+    while let Some(token) = ast_tokens.next() {
         match token {
             ast::HexToken::Byte(byte) => {
                 hir_tokens.push(hex_byte_hir_from_ast(byte));
             }
             ast::HexToken::NotByte(byte) => {
+                // ~?? is not allowed.
+                if byte.mask == 0 {
+                    return Err(Box::new(CompileError::invalid_pattern(
+                        ctx.report_builder,
+                        pattern_ident.name.to_string(),
+                        "negation of `??` is not allowed".to_string(),
+                        token.span().into(),
+                        None,
+                    )));
+                }
+
                 let class = match hex_byte_hir_from_ast(byte).into_kind() {
                     hir::HirKind::Class(mut class) => {
                         class.negate();
@@ -43,6 +56,7 @@ fn hex_tokens_hir_from_ast(
                     }
                     _ => unreachable!(),
                 };
+
                 hir_tokens.push(hir::Hir::class(class));
             }
             ast::HexToken::Alternative(alt) => {
@@ -54,21 +68,75 @@ fn hex_tokens_hir_from_ast(
                         ctx,
                         pattern_ident,
                         alt,
-                    ));
+                    )?);
                 }
 
                 hir_tokens.push(hir::Hir::alternation(alternatives))
             }
             ast::HexToken::Jump(jump) => {
-                if let Some(coalesced_span) = jump.coalesced_span {
+                let mut span = jump.span();
+                let mut jump = jump.clone();
+                let mut coalesced = false;
+
+                // Coalesce consecutive jumps into a single one. For example:
+                //
+                //  `[1-2][3-4]` becomes `[4-6]`
+                //  `[0-2][5-]` becomes `[5-]`
+                //  `[4][0-7]`  becomes `[4-11]`
+                //
+                while let Some(ast::HexToken::Jump(next)) = ast_tokens.peek() {
+                    match (jump.start, next.start) {
+                        (Some(s1), Some(s2)) => jump.start = Some(s1 + s2),
+                        (Some(s1), None) => jump.start = Some(s1),
+                        (None, Some(s2)) => jump.start = Some(s2),
+                        (None, None) => jump.start = None,
+                    }
+                    match (jump.end, next.end) {
+                        (Some(e1), Some(e2)) => jump.end = Some(e1 + e2),
+                        (_, _) => jump.end = None,
+                    }
+                    span = span.combine(&next.span());
+                    ast_tokens.next();
+                    coalesced = true;
+                }
+
+                if coalesced {
                     ctx.warnings.add(|| {
                         Warning::consecutive_jumps(
                             ctx.report_builder,
                             pattern_ident.name.to_string(),
-                            format!("{}", jump),
-                            coalesced_span,
+                            format!("{jump}"),
+                            (&span).into(),
                         )
                     });
+                }
+
+                match (jump.start, jump.end) {
+                    (Some(0), Some(0)) => {
+                        return Err(Box::new(CompileError::invalid_pattern(
+                            ctx.report_builder,
+                            pattern_ident.name.to_string(),
+                            "zero-length jumps are useless, remove it"
+                                .to_string(),
+                            span.into(),
+                            None,
+                        )));
+                    }
+                    (Some(start), Some(end)) if start > end => {
+                        return Err(Box::new(CompileError::invalid_pattern(
+                            ctx.report_builder,
+                            pattern_ident.name.to_string(),
+                            format!(
+                                "lower bound ({start}) is greater than upper bound ({end})"),
+                            span.into(),
+                            if coalesced {
+                                Some("consecutive jumps were coalesced into a single one".to_string())
+                            } else {
+                                None
+                            }
+                        )));
+                    }
+                    _ => {}
                 }
 
                 hir_tokens.push(hir::Hir::repetition(hir::Repetition {
@@ -81,7 +149,7 @@ fn hex_tokens_hir_from_ast(
         }
     }
 
-    hir::Hir::concat(hir_tokens)
+    Ok(hir::Hir::concat(hir_tokens))
 }
 
 fn hex_byte_hir_from_ast(byte: &ast::HexByte) -> hir::Hir {
@@ -112,32 +180,29 @@ mod tests {
     use regex_syntax::hir::{
         Class, ClassBytes, ClassBytesRange, Dot, Hir, HirKind, Repetition,
     };
+    use yara_x_parser::ast;
 
     use yara_x_parser::ast::{
-        HexAlternative, HexByte, HexJump, HexPattern, HexToken, HexTokens,
-        Ident,
+        HexAlternative, HexJump, HexPattern, HexToken, HexTokens, Ident,
     };
-
-    use yara_x_parser::report::ReportBuilder;
 
     use super::hex_byte_to_class;
     use crate::compiler::context::{CompileContext, VarStack};
+    use crate::compiler::report::ReportBuilder;
     use crate::compiler::Warnings;
+    use crate::re::hir;
     use crate::re::hir::class_to_masked_byte;
     use crate::symbols::StackedSymbolTable;
 
     #[test]
     fn hex_byte_to_hir() {
-        let hir =
-            super::hex_byte_hir_from_ast(&HexByte { value: 0x00, mask: 0x00 });
+        let hir = super::hex_byte_hir_from_ast(&ast::HexByte::new(0x00, 0x00));
         assert_eq!(hir.to_string(), r"(?-u:[\x00-\xFF])");
 
-        let hir =
-            super::hex_byte_hir_from_ast(&HexByte { value: 0x10, mask: 0xf0 });
+        let hir = super::hex_byte_hir_from_ast(&ast::HexByte::new(0x10, 0xf0));
         assert_eq!(hir.to_string(), r"(?-u:[\x10-\x1F])");
 
-        let hir =
-            super::hex_byte_hir_from_ast(&HexByte { value: 0x02, mask: 0x0f });
+        let hir = super::hex_byte_hir_from_ast(&ast::HexByte::new(0x02, 0x0f));
         assert_eq!(
             hir.to_string(),
             r#"(?-u:[\x02\x12"2BRbr\x82\x92\xA2\xB2\xC2\xD2\xE2\xF2])"#
@@ -159,143 +224,137 @@ mod tests {
             current_rule_patterns: &mut rule_patterns,
             warnings: &mut warnings,
             vars: VarStack::new(),
+            for_of_depth: 0,
         };
 
         let mut pattern = HexPattern {
-            span: Default::default(),
-            identifier: Ident { span: Default::default(), name: "test_ident" },
+            identifier: Ident::new("test_ident"),
             tokens: HexTokens {
                 tokens: vec![
-                    HexToken::Byte(HexByte { value: b'a', mask: 0xff }),
-                    HexToken::Byte(HexByte { value: b'b', mask: 0xff }),
-                    HexToken::Byte(HexByte { value: b'c', mask: 0xff }),
+                    HexToken::Byte(ast::HexByte::new(b'a', 0xff)),
+                    HexToken::Byte(ast::HexByte::new(b'b', 0xff)),
+                    HexToken::Byte(ast::HexByte::new(b'c', 0xff)),
                 ],
             },
-            modifiers: Default::default(),
+            ..Default::default()
         };
 
         assert_eq!(
             super::hex_pattern_hir_from_ast(&mut ctx, &pattern),
-            Hir::literal("abc".as_bytes())
+            Ok(Hir::literal("abc".as_bytes()))
         );
 
         pattern.tokens = HexTokens {
             tokens: vec![
-                HexToken::Byte(HexByte { value: 0x01, mask: 0xff }),
-                HexToken::Byte(HexByte { value: 0x02, mask: 0xff }),
-                HexToken::Byte(HexByte { value: 0x03, mask: 0xff }),
+                HexToken::Byte(ast::HexByte::new(0x01, 0xff)),
+                HexToken::Byte(ast::HexByte::new(0x02, 0xff)),
+                HexToken::Byte(ast::HexByte::new(0x03, 0xff)),
             ],
         };
 
         assert_eq!(
             super::hex_pattern_hir_from_ast(&mut ctx, &pattern),
-            Hir::literal([0x01, 0x02, 0x03])
+            Ok(Hir::literal([0x01, 0x02, 0x03]))
         );
 
         pattern.tokens = HexTokens {
             tokens: vec![
-                HexToken::Byte(HexByte { value: 0x01, mask: 0xff }),
-                HexToken::Byte(HexByte { value: 0x02, mask: 0xff }),
-                HexToken::Byte(HexByte { value: 0x03, mask: 0xff }),
-                HexToken::Byte(HexByte { value: 0x00, mask: 0x00 }),
-                HexToken::Byte(HexByte { value: 0x05, mask: 0xff }),
-                HexToken::Byte(HexByte { value: 0x06, mask: 0xff }),
+                HexToken::Byte(ast::HexByte::new(0x01, 0xff)),
+                HexToken::Byte(ast::HexByte::new(0x02, 0xff)),
+                HexToken::Byte(ast::HexByte::new(0x03, 0xff)),
+                HexToken::Byte(ast::HexByte::new(0x00, 0x00)),
+                HexToken::Byte(ast::HexByte::new(0x05, 0xff)),
+                HexToken::Byte(ast::HexByte::new(0x06, 0xff)),
             ],
         };
 
         assert_eq!(
             super::hex_pattern_hir_from_ast(&mut ctx, &pattern),
-            Hir::concat(vec![
+            Ok(Hir::concat(vec![
                 Hir::literal([0x01, 0x02, 0x03]),
                 Hir::dot(Dot::AnyByte),
                 Hir::literal([0x05, 0x06]),
-            ])
+            ]))
         );
 
         pattern.tokens = HexTokens {
             tokens: vec![
-                HexToken::Byte(HexByte { value: 0x01, mask: 0xff }),
-                HexToken::NotByte(HexByte { value: 0x02, mask: 0xff }),
-                HexToken::Byte(HexByte { value: 0x03, mask: 0xff }),
+                HexToken::Byte(ast::HexByte::new(0x01, 0xff)),
+                HexToken::NotByte(ast::HexByte::new(0x02, 0xff)),
+                HexToken::Byte(ast::HexByte::new(0x03, 0xff)),
             ],
         };
 
         assert_eq!(
             super::hex_pattern_hir_from_ast(&mut ctx, &pattern),
-            Hir::concat(vec![
+            Ok(Hir::concat(vec![
                 Hir::literal([0x01]),
                 Hir::class(Class::Bytes(ClassBytes::new(vec![
                     ClassBytesRange::new(0, 1),
                     ClassBytesRange::new(3, 255)
                 ]))),
                 Hir::literal([0x03]),
-            ])
+            ]))
         );
 
         pattern.tokens = HexTokens {
             tokens: vec![
-                HexToken::Byte(HexByte { value: 0x01, mask: 0xff }),
-                HexToken::NotByte(HexByte { value: 0x40, mask: 0xfe }),
-                HexToken::Byte(HexByte { value: 0x03, mask: 0xff }),
+                HexToken::Byte(ast::HexByte::new(0x01, 0xff)),
+                HexToken::NotByte(ast::HexByte::new(0x40, 0xfe)),
+                HexToken::Byte(ast::HexByte::new(0x03, 0xff)),
             ],
         };
 
         assert_eq!(
             super::hex_pattern_hir_from_ast(&mut ctx, &pattern),
-            Hir::concat(vec![
+            Ok(Hir::concat(vec![
                 Hir::literal([0x01]),
                 Hir::class(Class::Bytes(ClassBytes::new(vec![
                     ClassBytesRange::new(0, 0x3f),
                     ClassBytesRange::new(0x42, 0xff),
                 ]))),
                 Hir::literal([0x03]),
-            ])
+            ]))
         );
 
         pattern.tokens = HexTokens {
-            tokens: vec![HexToken::Alternative(Box::new(HexAlternative {
-                alternatives: vec![
+            tokens: vec![HexToken::Alternative(Box::new(
+                HexAlternative::new(vec![
                     HexTokens {
-                        tokens: vec![HexToken::Byte(HexByte {
-                            value: 0x01,
-                            mask: 0xff,
-                        })],
+                        tokens: vec![HexToken::Byte(ast::HexByte::new(
+                            0x01, 0xff,
+                        ))],
                     },
                     HexTokens {
-                        tokens: vec![HexToken::Byte(HexByte {
-                            value: 0x02,
-                            mask: 0xff,
-                        })],
+                        tokens: vec![HexToken::Byte(ast::HexByte::new(
+                            0x02, 0xff,
+                        ))],
                     },
-                ],
-            }))],
+                ]),
+            ))],
         };
-
         assert_eq!(
             super::hex_pattern_hir_from_ast(&mut ctx, &pattern),
-            Hir::alternation(
-                vec![Hir::literal([0x01]), Hir::literal([0x02]),]
-            )
+            Ok(Hir::alternation(vec![
+                Hir::literal([0x01]),
+                Hir::literal([0x02]),
+            ]))
         );
 
         pattern.tokens = HexTokens {
             tokens: vec![
-                HexToken::Byte(HexByte { value: 0x01, mask: 0xff }),
-                HexToken::Byte(HexByte { value: 0x02, mask: 0xff }),
-                HexToken::Byte(HexByte { value: 0x03, mask: 0xff }),
-                HexToken::Jump(HexJump {
-                    start: None,
-                    end: None,
-                    coalesced_span: None,
-                }),
-                HexToken::Byte(HexByte { value: 0x05, mask: 0xff }),
-                HexToken::Byte(HexByte { value: 0x06, mask: 0xff }),
+                HexToken::Byte(ast::HexByte::new(0x01, 0xff)),
+                HexToken::Byte(ast::HexByte::new(0x02, 0xff)),
+                HexToken::Byte(ast::HexByte::new(0x03, 0xff)),
+                HexToken::Jump(HexJump::new(None, None)),
+                HexToken::Byte(ast::HexByte::new(0x05, 0xff)),
+                HexToken::Byte(ast::HexByte::new(0x06, 0xff)),
             ],
         };
 
         assert_eq!(
             super::hex_pattern_hir_from_ast(&mut ctx, &pattern),
-            Hir::concat(vec![
+            Ok(Hir::concat(vec![
                 Hir::literal([0x01, 0x02, 0x03]),
                 Hir::repetition(Repetition {
                     min: 0,
@@ -304,34 +363,31 @@ mod tests {
                     sub: Box::new(Hir::dot(Dot::AnyByte)),
                 }),
                 Hir::literal([0x05, 0x06]),
-            ])
+            ]))
         );
     }
 
     #[test]
     fn class_to_hex() {
         assert_eq!(
-            class_to_masked_byte(&hex_byte_to_class(&HexByte {
-                value: 0x30,
-                mask: 0xF0
-            })),
-            Some(HexByte { value: 0x30, mask: 0xF0 })
+            class_to_masked_byte(&hex_byte_to_class(&ast::HexByte::new(
+                0x30, 0xff
+            ))),
+            Some(hir::HexByte { value: 0x30, mask: 0xff })
         );
 
         assert_eq!(
-            class_to_masked_byte(&hex_byte_to_class(&HexByte {
-                value: 0x05,
-                mask: 0x0F
-            })),
-            Some(HexByte { value: 0x05, mask: 0x0F })
+            class_to_masked_byte(&hex_byte_to_class(&ast::HexByte::new(
+                0x05, 0x0f
+            ))),
+            Some(hir::HexByte { value: 0x05, mask: 0x0f })
         );
 
         assert_eq!(
-            class_to_masked_byte(&hex_byte_to_class(&HexByte {
-                value: 0x08,
-                mask: 0xAA
-            })),
-            Some(HexByte { value: 0x08, mask: 0xAA })
+            class_to_masked_byte(&hex_byte_to_class(&ast::HexByte::new(
+                0x03, 0xff
+            ))),
+            Some(hir::HexByte { value: 0x03, mask: 0xff })
         );
 
         assert_eq!(
@@ -365,7 +421,7 @@ mod tests {
         {
             assert_eq!(
                 class_to_masked_byte(class),
-                Some(HexByte { value: 0x00, mask: 0x00 })
+                Some(hir::HexByte { value: 0x00, mask: 0x00 })
             );
         } else {
             unreachable!()
