@@ -1,20 +1,23 @@
 use pretty_assertions::assert_eq;
+use yara_x_parser::cst::SyntaxKind;
+use yara_x_parser::Parser;
 
 use crate::processor::{actions, Processor};
 use crate::tokens;
 use crate::tokens::Token::*;
 use crate::tokens::{categories, Token};
-use yara_x_parser::{GrammarRule, Parser};
 
-fn tokenize(source: &str) -> tokens::Tokens {
-    tokens::Tokens::new(Parser::new().build_cst(source).unwrap())
+fn tokenize(source: &str) -> Vec<Token> {
+    let events =
+        Parser::new(source.as_bytes()).into_cst_stream().whitespaces(false);
+    tokens::Tokens::new(events).collect()
 }
 
 #[test]
 fn no_rules() {
     // The output of a processor with no rules must be equal to the input.
     let input_tokens: Vec<Token> =
-        tokenize(r#"rule test { condition: true }"#).collect();
+        tokenize(r#"rule test { condition: true }"#);
 
     let processor = Processor::new(input_tokens.clone().into_iter());
     let output_tokens: Vec<Token> = processor.collect();
@@ -28,7 +31,7 @@ fn copy() {
     // always true and the action is actions::copy returns an output that is
     // exactly equal to the input.
     let input_tokens: Vec<Token> =
-        tokenize(r#"rule test { condition: true }"#).collect();
+        tokenize(r#"rule test { condition: true }"#);
 
     let processor = Processor::new(input_tokens.clone().into_iter())
         .add_rule(|_| true, actions::copy);
@@ -42,9 +45,10 @@ fn copy() {
 fn drop() {
     // Make sure that a processor with a single rule where the condition is
     // always true and the action is actions::drop doesn't return any tokens.
-    let mut processor =
-        Processor::new(tokenize(r#"rule test { condition: true }"#))
-            .add_rule(|_| true, actions::drop);
+    let mut processor = Processor::new(
+        tokenize(r#"rule test { condition: true }"#).into_iter(),
+    )
+    .add_rule(|_| true, actions::drop);
 
     assert!(processor.next().is_none());
 }
@@ -52,7 +56,7 @@ fn drop() {
 #[test]
 fn passthrough() {
     let input_tokens: Vec<Token> =
-        tokenize(r#"rule test { condition: true }"#).collect();
+        tokenize(r#"rule test { condition: true }"#);
 
     // This processor should remove all tokens except control tokens.
     let processor = Processor::new(input_tokens.clone().into_iter())
@@ -76,14 +80,16 @@ fn passthrough() {
     assert_eq!(
         output_tokens,
         vec![
-            Begin(GrammarRule::source_file),
-            Begin(GrammarRule::rule_decl),
-            Begin(GrammarRule::boolean_expr),
-            Begin(GrammarRule::boolean_term),
-            End(GrammarRule::boolean_term),
-            End(GrammarRule::boolean_expr),
-            End(GrammarRule::rule_decl),
-            End(GrammarRule::source_file),
+            Begin(SyntaxKind::SOURCE_FILE),
+            Begin(SyntaxKind::RULE_DECL),
+            Begin(SyntaxKind::CONDITION_BLK),
+            Begin(SyntaxKind::BOOLEAN_EXPR),
+            Begin(SyntaxKind::BOOLEAN_TERM),
+            End(SyntaxKind::BOOLEAN_TERM),
+            End(SyntaxKind::BOOLEAN_EXPR),
+            End(SyntaxKind::CONDITION_BLK),
+            End(SyntaxKind::RULE_DECL),
+            End(SyntaxKind::SOURCE_FILE),
         ]
     )
 }
@@ -91,14 +97,14 @@ fn passthrough() {
 #[test]
 fn swap() {
     let input_tokens =
-        vec![Keyword("foo"), Begin(GrammarRule::rule_decl), Keyword("bar")];
+        vec![Keyword(b"foo"), Begin(SyntaxKind::RULE_DECL), Keyword(b"bar")];
 
     let processor = Processor::new(input_tokens.clone().into_iter())
         .set_passthrough(*categories::CONTROL)
         .add_rule(
             |ctx| {
-                ctx.token(1) == &Keyword("foo")
-                    && ctx.token(2) == &Keyword("bar")
+                ctx.token(1) == &Keyword(b"foo")
+                    && ctx.token(2) == &Keyword(b"bar")
             },
             actions::swap(1, 2),
         );
@@ -107,18 +113,17 @@ fn swap() {
 
     assert_eq!(
         output_tokens,
-        vec![Keyword("bar"), Begin(GrammarRule::rule_decl), Keyword("foo")]
+        vec![Keyword(b"bar"), Begin(SyntaxKind::RULE_DECL), Keyword(b"foo")]
     )
 }
 
 #[test]
 fn drop_identifiers() {
     // Test a processor that drops only the tokens that are identifiers.
-    let input_tokens = tokens::Tokens::new(
-        Parser::new().build_cst(r#"rule test { condition: true }"#).unwrap(),
-    );
+    let input_tokens: Vec<Token> =
+        tokenize(r#"rule test { condition: true }"#);
 
-    let processor = Processor::new(input_tokens)
+    let processor = Processor::new(input_tokens.into_iter())
         // Drop identifiers.
         .add_rule(
             |ctx| ctx.token(1).is(*tokens::categories::IDENTIFIER),
@@ -130,22 +135,24 @@ fn drop_identifiers() {
     assert_eq!(
         output_tokens,
         vec![
-            Begin(GrammarRule::source_file),
-            Begin(GrammarRule::rule_decl),
-            Keyword("rule"),
+            Begin(SyntaxKind::SOURCE_FILE),
+            Begin(SyntaxKind::RULE_DECL),
+            Keyword(b"rule"),
             // This is the dropped identifier.
             // Identifier("test"),
-            Punctuation("{"),
-            Keyword("condition"),
-            Punctuation(":"),
-            Begin(GrammarRule::boolean_expr),
-            Begin(GrammarRule::boolean_term),
-            Keyword("true"),
-            End(GrammarRule::boolean_term),
-            End(GrammarRule::boolean_expr),
-            Punctuation("}"),
-            End(GrammarRule::rule_decl),
-            End(GrammarRule::source_file),
+            Punctuation(b"{"),
+            Begin(SyntaxKind::CONDITION_BLK),
+            Keyword(b"condition"),
+            Punctuation(b":"),
+            Begin(SyntaxKind::BOOLEAN_EXPR),
+            Begin(SyntaxKind::BOOLEAN_TERM),
+            Keyword(b"true"),
+            End(SyntaxKind::BOOLEAN_TERM),
+            End(SyntaxKind::BOOLEAN_EXPR),
+            End(SyntaxKind::CONDITION_BLK),
+            Punctuation(b"}"),
+            End(SyntaxKind::RULE_DECL),
+            End(SyntaxKind::SOURCE_FILE),
         ]
     )
 }
@@ -154,45 +161,47 @@ fn drop_identifiers() {
 fn insert_global() {
     // Test a processor that inserts a "global" keyword before "rule".
     use crate::tokens::Token::*;
-    use yara_x_parser::GrammarRule;
 
-    let processor =
-        Processor::new(tokenize(r#"rule test { condition: true }"#))
-            // Drop identifiers.
-            .add_rule(
-                |c| match (c.token(-1), c.token(1)) {
-                    // The "rule" keyword is already preceded by "global", do
-                    // nothing.
-                    (Keyword("global"), Keyword("rule")) => false,
-                    // In all other cases where "rule" is found insert "global".
-                    (_, Keyword("rule")) => true,
-                    // For all other tokens do nothing.
-                    _ => false,
-                },
-                actions::insert(Keyword("global")),
-            );
+    let processor = Processor::new(
+        tokenize(r#"rule test { condition: true }"#).into_iter(),
+    )
+    // Drop identifiers.
+    .add_rule(
+        |c| match (c.token(-1), c.token(1)) {
+            // The "rule" keyword is already preceded by "global", do
+            // nothing.
+            (Keyword(b"global"), Keyword(b"rule")) => false,
+            // In all other cases where "rule" is found insert "global".
+            (_, Keyword(b"rule")) => true,
+            // For all other tokens do nothing.
+            _ => false,
+        },
+        actions::insert(Keyword(b"global")),
+    );
 
     let output_tokens: Vec<Token> = processor.collect();
 
     assert_eq!(
         output_tokens,
         vec![
-            Begin(GrammarRule::source_file),
-            Begin(GrammarRule::rule_decl),
-            Keyword("global"),
-            Keyword("rule"),
-            Identifier("test"),
-            Punctuation("{"),
-            Keyword("condition"),
-            Punctuation(":"),
-            Begin(GrammarRule::boolean_expr),
-            Begin(GrammarRule::boolean_term),
-            Keyword("true"),
-            End(GrammarRule::boolean_term),
-            End(GrammarRule::boolean_expr),
-            Punctuation("}"),
-            End(GrammarRule::rule_decl),
-            End(GrammarRule::source_file),
+            Begin(SyntaxKind::SOURCE_FILE),
+            Begin(SyntaxKind::RULE_DECL),
+            Keyword(b"global"),
+            Keyword(b"rule"),
+            Identifier(b"test"),
+            Punctuation(b"{"),
+            Begin(SyntaxKind::CONDITION_BLK),
+            Keyword(b"condition"),
+            Punctuation(b":"),
+            Begin(SyntaxKind::BOOLEAN_EXPR),
+            Begin(SyntaxKind::BOOLEAN_TERM),
+            Keyword(b"true"),
+            End(SyntaxKind::BOOLEAN_TERM),
+            End(SyntaxKind::BOOLEAN_EXPR),
+            End(SyntaxKind::CONDITION_BLK),
+            Punctuation(b"}"),
+            End(SyntaxKind::RULE_DECL),
+            End(SyntaxKind::SOURCE_FILE),
         ]
     )
 }
@@ -201,49 +210,52 @@ fn insert_global() {
 fn in_rule() {
     // Test a processor that inserts a "global" keyword before "rule".
     use crate::tokens::Token::*;
-    use yara_x_parser::GrammarRule;
 
-    let tokens = Processor::new(tokenize(r#"rule test { condition: true }"#))
-        .add_rule(
-            |c| {
-                c.in_rule(GrammarRule::boolean_expr, false)
-                    && c.token(-1)
-                        .neq(&Literal("<next token is in boolean_expr>"))
-            },
-            actions::insert(Literal("<next token is in boolean_expr>")),
-        )
-        .add_rule(
-            |c| {
-                c.in_rule(GrammarRule::boolean_term, false)
-                    && c.token(-1)
-                        .neq(&Literal("<next token is in boolean_term>"))
-            },
-            actions::insert(Literal("<next token is in boolean_term>")),
-        );
+    let tokens = Processor::new(
+        tokenize(r#"rule test { condition: true }"#).into_iter(),
+    )
+    .add_rule(
+        |c| {
+            c.in_rule(SyntaxKind::BOOLEAN_EXPR, false)
+                && c.token(-1)
+                    .neq(&Literal(b"<next token is in boolean_expr>"))
+        },
+        actions::insert(Literal(b"<next token is in boolean_expr>")),
+    )
+    .add_rule(
+        |c| {
+            c.in_rule(SyntaxKind::BOOLEAN_TERM, false)
+                && c.token(-1)
+                    .neq(&Literal(b"<next token is in boolean_term>"))
+        },
+        actions::insert(Literal(b"<next token is in boolean_term>")),
+    );
 
     let output_tokens: Vec<Token> = tokens.collect();
 
     assert_eq!(
         output_tokens,
         vec![
-            Begin(GrammarRule::source_file),
-            Begin(GrammarRule::rule_decl),
-            Keyword("rule"),
-            Identifier("test"),
-            Punctuation("{"),
-            Keyword("condition"),
-            Punctuation(":"),
-            Begin(GrammarRule::boolean_expr),
-            Literal("<next token is in boolean_expr>"),
-            Begin(GrammarRule::boolean_term),
-            Literal("<next token is in boolean_term>"),
-            Keyword("true"),
-            Literal("<next token is in boolean_expr>"),
-            End(GrammarRule::boolean_term),
-            End(GrammarRule::boolean_expr),
-            Punctuation("}"),
-            End(GrammarRule::rule_decl),
-            End(GrammarRule::source_file),
+            Begin(SyntaxKind::SOURCE_FILE),
+            Begin(SyntaxKind::RULE_DECL),
+            Keyword(b"rule"),
+            Identifier(b"test"),
+            Punctuation(b"{"),
+            Begin(SyntaxKind::CONDITION_BLK),
+            Keyword(b"condition"),
+            Punctuation(b":"),
+            Begin(SyntaxKind::BOOLEAN_EXPR),
+            Literal(b"<next token is in boolean_expr>"),
+            Begin(SyntaxKind::BOOLEAN_TERM),
+            Literal(b"<next token is in boolean_term>"),
+            Keyword(b"true"),
+            Literal(b"<next token is in boolean_expr>"),
+            End(SyntaxKind::BOOLEAN_TERM),
+            End(SyntaxKind::BOOLEAN_EXPR),
+            End(SyntaxKind::CONDITION_BLK),
+            Punctuation(b"}"),
+            End(SyntaxKind::RULE_DECL),
+            End(SyntaxKind::SOURCE_FILE),
         ]
     )
 }
